requirements.txt
anthropic>=0.40.0
apify-client>=1.7.0
gspread>=6.0.0
google-auth>=2.28.0
openpyxl>=3.1.2
pyyaml>=6.0.1
requests>=2.31.0
